Fix remaining low/informational bugs from code audit
Mace4: use Current_depth global for checkpoint instead of fragile
stack-scanning heuristic (msearch.c).

Mace4: use long long for eterm/nterm ID computation to prevent
intermediate integer overflow (propagate.c, negprop.c).

Mace4: check_parent_pointers now recurses correctly instead of
calling set_parent_pointers (ground.c, debug-only function).

Mace4: %d -> %u format for unsigned cross_offs field (print.c).

Mace4: next_domain_size loop exits when n exceeds end_size instead
of potentially looping forever (msearch.c).

TPTP: warn on stderr when token is truncated at 4096 chars
(tptp_parse.c).

+++ mace4.src/negprop.c
@@ -46,15 +46,15 @@ extern struct mace_stats Mstats;
 static
 int nterm_id(Term t)
 {
-  int id = Sn_to_mace_sn[SYMNUM(t)]->base;
-  int mult = 1;
+  long long id = Sn_to_mace_sn[SYMNUM(t)]->base;
+  long long mult = 1;
   int i;
   for (i = ARITY(t)-1; i >= 0; i--) {
     if (VARIABLE(ARG(t,i)))
       id += VARNUM(ARG(t,i)) * mult;
     mult *= Domain_size;
   }
-  return id;
+  return (int) id;
 }  /* nterm_id */
 
 /*************
